requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
banks==2.4.1
bcrypt==5.0.0
build==1.4.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.10
filelock==3.29.0
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
greenlet==3.4.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
llama-index==0.14.21
llama-index-core==0.14.21
llama-index-embeddings-ollama==0.9.0
llama-index-embeddings-openai==0.6.0
llama-index-instrumentation==0.5.0
llama-index-llms-ollama==0.10.1
llama-index-llms-openai==0.7.5
llama-index-workflows==2.19.1
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.25.0
openai==2.32.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
overrides==7.7.0
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
pypdf==6.10.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.1
